Safari-609.1.11.1/LayoutTests/media/video-play-pause-exception-expected.txt


Video has no src. Test that the playing event is not dispatched.

RUN(handlePromise(video.play()))
RUN(video.pause())
EVENT(waiting)
EVENT(pause)
EXPECTED (video.networkState == '0') OK
END OF TEST
